add back depwarn option if invokelatest succeeded
Also remove `isdefined` shim, as that feels less logical to me with the
current implementation that uses invokelatest. If the implementation was
changed to use `invokenext`, I think it would make sense to add that back
and also to update `names` to return these also. But invoke next
(similar to backdated constants and #59735) prints an incorrect
suggestion of using invokelatest to get the same behavior. That message
is correct with this current commit state, but wrong relative to the
current implementations that use backdated.