Allow globalref on lhs of op= (#59762)
Fix https://github.com/JuliaLang/julia/issues/59008. `globalref`s are OK
to assign to.

This check isn't actually necessary since we check the lhs again when
re-expanding the assignment, and most of the forms allowed through
aren't valid with a non-dotted `op=`, but deleting it makes some error
messages worse.