Align interpreter and codegen error behavior of setglobal! and friends (#59766)
Currently this is an ErrorException in the runtime/interpreter, but a
TypeError in codegen. This is not permitted - which error is thrown is
semantically observable and codegen is not permitted to change it.
Worse, inference is also inconsistent about whether this is TypeError or
ErrorException, so this could actually lead to type confusion and
crashes. Fix all that by having the runtime also emit a TypeError here.
However, in order to not lose the binding name in the error message,
adjust the TypeError context field to permit a binding.