Implement setindex for ImmutableDict
This adds a method of `Base.setindex` for `Base.ImmutableDict`, as suggested in
JuliaObjects/Accessors.jl#216.

For a key that is already in the `ImmutableDict`, rather than add it again and
shadow it, as described in the `ImmutableDict` doc string, I opted to exclude
the original key and then add in the new one.