Small change to docs for module import syntax (#54575)
~~Note: the base branch (target branch) of this PR is `kc/warn_using`
(#42080), NOT `master`.~~

The base branch of this PR used to be `kc/warn_using` (#42080), but now
it's `master`.

---

This small change was originally part of #53428. However, it caused
confusion during PR review. So I removed this change from #53428 (in
https://github.com/JuliaLang/julia/pull/53428/commits/862a5466ba19acee25d566488431123fe5494ca1),
and I've put that change into this PR.

Just to clarify, the language about "keeping the namespace clean"
already existed on master prior to #53428 (and prior to this PR). All
this PR does is:
1. Move the existing "clean namespace" sentence down.
2. In that existing sentence, change "`import ModuleName`" to "`import
ModuleName` or `using ModuleName: ModuleName`".

Co-authored-by: Viral B. Shah <viral@juliacomputing.com>