Add single-argument `foreach` method to consume iterable
We seem to lack a way to just consume an iterable with a function.

This is not a big deal, but a little quality-of-life improvement and seems to be more or less standard in programming languages nowadays.

`foreach` seems to be the most fitting name for Julia according to this overview from ChatGPT 5:
Summary (just the primary names/keywords):
- Python: for, deque(...)
- JavaScript: for...of
- Java: forEachRemaining, forEach
- C++: for_each (std / ranges)
- C#: foreach
- Go: for range
- Ruby: each
- Rust: for_each
- Swift: forEach
- Kotlin: forEach

Implemented with help from ChatGPT 5.