precompile: fix error / output printing
I counted at least 4 different places where an (overlapping) subset of
package IO might be printed. Now try to be slightly more careful about
that: progress goes to logio (either io or devnull) while output goes to
io always.

Additionally, we don't return an extra copy of some output in the error
message (since we already printed that).

Additionally, we print all output if the requested packages failed, but
we print only output from successful package if all requested packages
succeeded. If a package was successful, then its output is required to
be printed since it will not be regenerated in the future (we don't
reprint precompile output on loading since that might be obnoxious for
us to implement). If a required package is successful however, none of
the failures mattered and they will be regenerated in the future when
they are required, so those should be filtered out.