Use libuv in absrealpath to fix Korean path names on Windows (#59910)
Fixes #33486

The issue in #33486 was caused by the use of `GetFullPathName` on
Windows in the `absrealpath` C function.

`GetFullPathName` does not always return a UTF-8 encoded path.
`absrealpath` is used to set the paths in `jl_options`.
So if Julia is run from a non-ASCII directory with a non-UTF-8 locale,
the paths in `jl_options` are broken.

This PR fixes this by using `uv_fs_realpath` and `uv_cwd` from libuv,
which handles conversion of Windows paths to WTF-8.