cgutils: fix write barrier of atomic-setonce
Thanks to PermutationGroups.jl for dog-fooding our typos! This now
matches the runtime implementation in datatype.c (setonce_bits) and
ensures that we properly compute the success bit. The test that was
intended to check for this is now written to inline that struct and thus
actually test for this (undef fields are not eligible for inlining).

Fix #59883