Add superscript capital C, F, Q (#60007)
Is this correct? They look a bit off on GitHub, but they match what
Wikipedia is saying in the first line of this table:
https://en.wikipedia.org/wiki/Unicode_subscripts_and_superscripts#Latin,_Greek,_Cyrillic,_and_IPA_tables.

They also seem coherent with the other capital superscripts, e.g. here
is B (old): https://codepoints.net/U+1d2e and C (new in this PR):
https://codepoints.net/U+A7F2.