Make some tests not fail with uid 0
When running as UID 0 (system root, but also container root in various
sandboxing technologies), the system ignores unix permissions. We
have a number of tests that assume that a 000 mode means the file
is not readable. Adjust these tests appropriately.

There are more failures, but I'll look at those in a follow-up.