libgit2/test: Handle a partially broken openssl install
The path to the openssl config file gets baked in at build time
and is wrong for the openssl we ship in JLLs causing trouble here
if there is no systemwide openssl. However, for our purposes an
empty openssl cnf is fine and probably better for reproducability
anyway, so just set that.