Uniqueness `gensym` guarantee goes beyond modules.
I am suggesting this after zulip thread [there](https://julialang.zulipchat.com/#narrow/channel/137791-general/topic/gensym.20scope/with/553881361). Thank you @adienes @vchuravy @MasonProtter for discussion.

The guarantee offered by `gensym` is stronger than just *"within the current module"* because it actually spans the entire session.

Because of a bug outlined in #44903, this unfortunately does not mean that uniqueness is guaranteed within a module yet. But I believe it is still the intent.