trimming: fix-up support for `Core._apply_iterate`
We only support a few built-in versions of this call (due to
https://github.com/JuliaLang/julia/issues/57830), but those should
at least be working now with this small tweak.

This required a fix-up for built-in functions to be included in the
method table for `--trim`'d executables, so that we can dynamically
dispatch to them, as the verifier expects.