Markdown: don't allow space between parts of a link (#59977)
In Markdown, links are written as `[desc](URL)`, and no
space is allowed between `]` and `(`. However, Julia's
Markdown parser explicitly accepted it (not clear why).

This may break links in a few package manuals, but the fix
will be easy in each case (just remove the extra spaces).

But being more consistent with most (all??) other Markdown
implementations seems more important to me; it is very
surprising when one copies basic Markdown that works fine
elsewhere into a Julia package manual and it suddenly is
treated completely differently.

Resolves JuliaDocs/Documenter.jl#2681