Move `.PHONY` statements in `Makefile`s and add missing ones (#60075)
This moves most `.PHONY` statements in our `Makefile`s next to the
target they are marking as phony. This then makes it easy to spot those
targets which clearly should be marked as phony but aren't. Subsequently
this is then resolved.

Also some `.PHONY` statements for targets that do not exist anymore are
also removed.