Fix windows profiler deadlock (#60056)
Wasn't sure if we could get away with deleting the `uv_thread_detach`,
but
apparently we can. On Windows, it's definitely unsafe to keep using the
thread
handle after closing it. POSIX is a little less clear about what you're
allowed
to do with a `pthread_t` after detaching it, but the GC threads never
exit
normally anyway.

Fixes #60042.