allow finding stdlibs that are normal packages in the manifest but are now stdlibs (#60061)
This helps handling manifest from earlier julia versions. This is kind
of the `locate` version of the `identify` fallback in
https://github.com/JuliaLang/julia/pull/56148.

Need to write a test when this happens (and verify that this fix works).

Noted in a slack #gripes comment.

---------

Co-authored-by: KristofferC <kristoffer.carlsson@juliacomputing.com>
Co-authored-by: Max Horn <max@quendi.de>
(cherry picked from commit 6fddac850a00441cc5d3833cc3c441132cd49a5f)