Use hand-written PLT thunks
Unfortunately the "portable" LLVM way of generating thunks doesn't generate the
code we want.  Instead, on platforms where it makes sense, we'll steal the LLD
PLT thunk code, but in disassembled form.  At some point this should be moved to
after linking, where it can be in assembled form again.  Amusingly it will be
more portable in assembled form, because the assembler syntax for relocations
differs between object formats.